custom_toolkit/pyproject.toml
[tool.poetry]
name = "arcade_custom_toolkit"
version = "0.1.0"
description = "Custom toolkit developed to get the hang of Arcade AI"
authors = ["Daniel Arvelini Nordhausen <daniel@arvelini.com>"]

[tool.poetry.dependencies]
python = "^3.10"
arcade-ai = "~0.1.3"

[tool.poetry.dev-dependencies]
pytest = "^8.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"